requirements/base.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.20
colorific==0.3
colormath==2.1.1
diff-match-patch==20121119
dj-database-url==0.3.0
dj.choices==0.10.0
django-appconf==1.0.1
django-braces==1.8.1
django-celery==3.1.17
django-classy-tags==0.6.2
django-compressor==1.4 # found around fabfile.py:602 - should be here
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-downloadview==1.8
django-extensions==1.6.1
django-filer==1.1.1
django-filter==0.12.0
django-geojson==2.8.1
django-health-check==1.0.2
django-import-export==0.4.2
django-leaflet==0.18.0
django-mptt==0.8.0
django-oauth-toolkit==0.10.0
django-object-actions==0.7.0
django-orderable==4.0.2
django-overextends==0.4.1
django-picklefield==0.3.2
django-polymorphic==0.8.1
django-rest-swagger==0.3.5
django-reversion==1.10.1
django-rosetta==0.7.6
django-suit==0.2.16
django-taggit==0.18.0
django-taggit-autosuggest==0.2.8
django-wysiwyg==0.7.1
django==1.8.6
djangorestframework==3.3.2
djangorestframework-jsonp==1.0.2
djangorestframework-xml==1.3.0
easy-thumbnails==2.3
elasticsearch-dsl==0.0.8
elasticsearch==1.6.0
elasticutils==0.10.3
#enum34==1.0.4
eventlet==0.18.1
flower==0.8.4
geojson==1.3.1
gunicorn==19.4.5
#gnureadline==6.3.3
html5lib==0.9999999
isodate==0.5.4
jsonfield==1.0.3
jsonpickle==0.9.2
kombu==3.0.33
lxml==3.5.0
networkx==1.10
path.py==8.1.2
numpy==1.10.4
pillow==3.1.0
polib==1.0.7
psycopg2==2.6.1
py==1.4.31
pyparsing==2.0.7
pyproj==1.9.5.1
python-dateutil==2.4.2
python3-memcached==1.51 # found around fabfile.py:602 - should be here
pytz==2015.7
pyyaml==3.11
raven==5.10.1
rdflib-jsonld==0.3
rdflib==4.2.1
requests==2.9.1
retry==0.9.1
setproctitle==1.1.8 # found around fabfile.py:602 - should be here
six==1.10.0
sparqlwrapper==1.7.6
sqlparse==0.1.18
tablib==0.10.0
tzlocal==1.2
unidecode==0.4.19
urllib3==1.14
wagtail==1.3.1
webcolors==1.5
wheel==0.26.0
willow==0.2.2
